requirements.txt
# smart_food_bot/requirements.txt
# Core
torch>=2.0.0
transformers>=4.40.0
ujson>=5.9.0
scikit-learn>=1.3.0
rank-bm25>=0.2.2
faiss-cpu>=1.8.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pymongo>=4.6.0

# Optional/quality
numpy>=1.24.0
